pyproject.toml
[tool.poetry]
name = "nonebot-plugin-fornite"
version = "0.0.1"
description = "堡垒之夜插件"
authors = ["fllesser <fllessive@gmail.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_fortnite"}]
repository = "https://github.com/fllesser/nonebot-plugin-fornite"
keywords = ["nonebot", "nonebot2"]


[tool.poetry.dependencies]
python = "^3.10"
fortnite-api = "^3.2.1"
nonebot2 = "^2.4.0"
nonebot-adapter-onebot = "^2.4.6"
nonebot-plugin-uninfo = "^0.6.5"
nonebot-plugin-alconna = "^0.54.2"
nonebot-plugin-localstore = "^0.7.3"
nonebot-plugin-apscheduler = "^0.5.0"

[tool.poetry.group.test.dependencies]
nb-cli = "^1.4.2"
nonebot2 = {version = "^2.4.0", extras = ["fastapi"]}

[tool.nonebot]
adapters = [{name="Onebot", module_name="nonebot.adapters.onebot.v11"}]
plugins = ["nonebot_plugin_fortnite"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"